+++ src/history/fat_map.rs
@@ -54,9 +54,9 @@ impl FatMapHistorySource {
 }
 
 impl HistorySource for FatMapHistorySource {
-    fn new(input_size: usize, max_order: usize) -> FatMapHistorySource {
+    fn new(max_window_size: usize, max_order: usize) -> FatMapHistorySource {
         FatMapHistorySource {
-            input: Vec::with_capacity(input_size),
+            input: Vec::with_capacity(max_window_size),
             input_cursor: 0,
             bit_index: 7,
             max_order,
@@ -67,7 +67,8 @@ impl HistorySource for FatMapHistorySource {
     fn start_new_byte(&mut self) {
         assert_eq!(self.bit_index, 7);
         assert_eq!(self.input_cursor, self.input.len());
-        assert_ne!(self.input.len(), self.input.capacity());
+        assert_ne!(self.input.len(), self.input.capacity(),
+                   "input window is filled up, but sliding is not implemented");
         self.input.push(0);
     }